+++ livre/lib/cache/DataObject.h
@@ -1,5 +1,6 @@
-/* Copyright (c) 2011-2014, EPFL/Blue Brain Project
+/* Copyright (c) 2011-2015, EPFL/Blue Brain Project
  *                     Ahmet Bilgili <ahmet.bilgili@epfl.ch>
+ *                     Daniel Nachbaur <daniel.nachbaur@epfl.ch>
  *
  * This file is part of Livre <https://github.com/BlueBrain/Livre>
  *
@@ -17,49 +18,46 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#ifndef _TextureObject_h_
-#define _TextureObject_h_
+#ifndef _DataObject_h_
+#define _DataObject_h_
 
-#include <livre/lib/api.h>
 #include <livre/lib/types.h>
-#include <livre/core/mathTypes.h>
-#include <livre/core/cache/CacheObject.h>
+
+#include <livre/core/cache/CacheObject.h> // base class
 
 namespace livre
 {
 
 /**
- * The TextureObject class holds the informarmation for the data which is on the GPU.
-  */
-class TextureObject : public CacheObject
+ * The DataObject class stores raw data from the volume data source.
+ */
+class DataObject : public CacheObject
 {
-    friend class TextureCache;
-
 public:
+    ~DataObject();
 
-    LIVRE_API virtual ~TextureObject();
-
-    /**
-     * @return The texture state.
-     */
-    LIVRE_API TextureStatePtr getTextureState();
-
-    /**
-     * @return The texture state ( const ).
-     */
-    LIVRE_API ConstTextureStatePtr getTextureState() const;
+    /** @return A pointer to the data or 0 if no data is loaded. */
+    const void* getDataPtr() const;
 
 private:
+    friend class DataCache;
 
-    TextureObject( const CacheId& cacheId,
-                   TextureCache& textureCache );
+    /**
+     * Constructor
+     * @param cacheId is the unique identifier
+     * @param dataCache the histogram source data is retrieved from data cache
+     * @throws std::runtime_error when the data cache does not have the data for cache id
+     */
+    DataObject( const CacheId& cacheId, DataCache& dataCache );
 
+    /** @copydoc livre::CacheObject::getSize */
     size_t _getSize() const final;
 
     struct Impl;
     std::unique_ptr<Impl> _impl;
+
 };
 
 }
 
-#endif // _TextureObject_h_
+#endif // _DataObject_h_